moved column logic into useColumns
